refactor: update import path for HarEntry and simplify pattern match usage
